Modular and multi-file development
