+++ data/network_traffic.csv
@@ -1,101 +1,19 @@
 timestamp,src_ip,dst_ip,protocol,length
-1718896522.263999,10.157.152.159,157.148.45.163,TCP,918
-1718896522.264076,10.157.152.159,157.148.45.163,TCP,1782
-1718896522.264105,10.157.152.159,157.148.45.163,TCP,1734
-1718896522.285271,157.148.45.163,10.157.152.159,TCP,60
-1718896522.285271,157.148.45.163,10.157.152.159,TCP,60
-1718896522.285271,157.148.45.163,10.157.152.159,TCP,60
-1718896522.285318,10.157.152.159,157.148.45.163,TCP,1350
-1718896522.302325,157.148.45.163,10.157.152.159,TCP,60
-1718896524.264765,10.157.152.159,157.148.45.163,TCP,454
-1718896524.285437,157.148.45.163,10.157.152.159,TCP,60
-1718896526.577321,10.157.152.159,157.148.45.225,TCP,566
-1718896526.60045,157.148.45.225,10.157.152.159,TCP,60
-1718896526.632578,157.148.45.225,10.157.152.159,TCP,453
-1718896526.687741,10.157.152.159,157.148.45.225,TCP,54
-1718896528.266743,10.157.152.159,157.148.45.163,TCP,471
-1718896528.284462,157.148.45.163,10.157.152.159,TCP,60
-1718896528.284509,10.157.152.159,157.148.45.163,TCP,488
-1718896528.318419,157.148.45.163,10.157.152.159,TCP,60
-1718896528.700868,120.232.239.80,10.157.152.159,TCP,60
-1718896528.700935,10.157.152.159,120.232.239.80,TCP,54
-1718896528.701028,10.157.152.159,120.232.239.80,TCP,78
-1718896528.701113,10.157.152.159,120.232.239.80,TCP,54
-1718896528.731815,10.157.152.159,192.168.254.226,UDP,88
-1718896528.733932,192.168.254.226,10.157.152.159,UDP,466
-1718896528.734606,10.157.152.159,120.241.118.6,TCP,66
-1718896528.742908,120.232.239.80,10.157.152.159,TCP,60
-1718896528.742908,120.232.239.80,10.157.152.159,TCP,60
-1718896528.777197,120.241.118.6,10.157.152.159,TCP,66
-1718896528.777309,10.157.152.159,120.241.118.6,TCP,54
-1718896528.777613,10.157.152.159,120.241.118.6,TCP,336
-1718896528.824294,120.241.118.6,10.157.152.159,TCP,60
-1718896529.012608,10.157.152.159,20.198.162.76,TCP,97
-1718896529.02834,120.241.118.6,10.157.152.159,TCP,1494
-1718896529.02834,120.241.118.6,10.157.152.159,TCP,1351
-1718896529.028506,10.157.152.159,120.241.118.6,TCP,54
-1718896529.031198,10.157.152.159,120.241.118.6,TCP,118
-1718896529.031249,10.157.152.159,120.241.118.6,TCP,169
-1718896529.031343,10.157.152.159,120.241.118.6,TCP,450
-1718896529.067519,120.241.118.6,10.157.152.159,TCP,1351
-1718896529.067519,120.241.118.6,10.157.152.159,TCP,60
-1718896529.067519,120.241.118.6,10.157.152.159,TCP,60
-1718896529.067519,120.241.118.6,10.157.152.159,TCP,60
-1718896529.06755,10.157.152.159,120.241.118.6,TCP,66
-1718896529.093243,20.198.162.76,10.157.152.159,TCP,228
-1718896529.14003,10.157.152.159,20.198.162.76,TCP,54
-1718896529.153145,120.241.118.6,10.157.152.159,TCP,78
-1718896529.153145,120.241.118.6,10.157.152.159,TCP,60
-1718896529.15322,10.157.152.159,120.241.118.6,TCP,54
-1718896529.153307,10.157.152.159,120.241.118.6,TCP,78
-1718896529.153373,10.157.152.159,120.241.118.6,TCP,54
-1718896529.197484,120.241.118.6,10.157.152.159,TCP,60
-1718896529.197484,120.241.118.6,10.157.152.159,TCP,60
-1718896531.105253,10.157.152.159,192.168.254.226,UDP,95
-1718896531.107261,192.168.254.226,10.157.152.159,UDP,275
-1718896531.107666,10.157.152.159,220.202.155.172,TCP,66
-1718896531.115897,220.202.155.172,10.157.152.159,TCP,66
-1718896531.116003,10.157.152.159,220.202.155.172,TCP,54
-1718896531.116314,10.157.152.159,220.202.155.172,TCP,453
-1718896531.125048,220.202.155.172,10.157.152.159,TCP,60
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1264
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143677,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.143833,10.157.152.159,220.202.155.172,TCP,66
-1718896531.143892,10.157.152.159,220.202.155.172,TCP,66
-1718896531.143909,10.157.152.159,220.202.155.172,TCP,74
-1718896531.143931,10.157.152.159,220.202.155.172,TCP,66
-1718896531.143945,10.157.152.159,220.202.155.172,TCP,74
-1718896531.143968,10.157.152.159,220.202.155.172,TCP,66
-1718896531.143993,10.157.152.159,220.202.155.172,TCP,66
-1718896531.144018,10.157.152.159,220.202.155.172,TCP,66
-1718896531.144045,10.157.152.159,220.202.155.172,TCP,54
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
-1718896531.144187,220.202.155.172,10.157.152.159,TCP,1494
+1718918758.989301,172.20.10.2,183.239.129.229,TCP,120
+1718918758.996796,172.20.10.1,224.0.0.251,UDP,100
+1718918759.179437,172.20.10.2,120.232.239.9,TCP,55
+1718918759.184559,183.239.129.229,172.20.10.2,TCP,66
+1718918759.184559,183.239.129.229,172.20.10.2,TCP,251
+1718918759.227434,172.20.10.2,183.239.129.229,TCP,54
+1718918759.23105,120.232.239.9,172.20.10.2,TCP,66
+1718918759.824687,172.20.10.2,183.239.129.229,TCP,55
+1718918762.387535,172.20.10.2,20.187.186.89,TCP,55
+1718918762.891606,172.20.10.1,224.0.0.251,UDP,100
+1718918762.891606,20.187.186.89,172.20.10.2,TCP,66
+1718918763.198977,172.20.10.2,36.155.202.119,TCP,454
+1718918763.248481,36.155.202.119,172.20.10.2,TCP,54
+1718918763.248539,172.20.10.2,36.155.202.119,TCP,1342
+1718918763.287509,36.155.202.119,172.20.10.2,TCP,54
+1718918764.78357,172.20.10.2,157.148.45.91,TCP,566
+1718918765.066678,172.20.10.2,157.148.45.91,TCP,566
+1718918765.199745,172.20.10.2,36.155.202.119,TCP,500
